cloud compute bugs, generate estimate, before helper functions
